qc/outputs: new submodule with functions to predict and check QC outputs.
Adds new qc/outputs.py submodule which collects together functions for
predicting and checking QC outputs from qc/illumina_qc.py (also removed as
part of this commit) and qc/pipeline.py.
