--- auto_process_ngs/qc/illumina_qc.py
@@ -1,92 +1,0 @@
-#!/usr/bin/env python
-#
-#     illumina_qc: running and validating QC
-#     Copyright (C) University of Manchester 2018-2019 Peter Briggs
-#
-"""
-Provides utility functions for QC outputs.
-
-Provides the following functions:
-
-- fastq_screen_output: get names for fastq_screen outputs
-- fastqc_output: get names for FastQC outputs
-- fastq_strand_output: get name for fastq_strand.py output
-"""
-
-#######################################################################
-# Imports
-#######################################################################
-
-import os
-import logging
-from bcftbx.qc.report import strip_ngs_extensions
-
-# Module specific logger
-logger = logging.getLogger(__name__)
-
-#######################################################################
-# Functions
-#######################################################################
-
-def fastq_screen_output(fastq,screen_name):
-    """
-    Generate name of fastq_screen output files
-
-    Given a Fastq file name and a screen name, the outputs from
-    fastq_screen will look like:
-
-    - {FASTQ}_{SCREEN_NAME}_screen.png
-    - {FASTQ}_{SCREEN_NAME}_screen.txt
-
-    Arguments:
-       fastq (str): name of Fastq file
-       screen_name (str): name of screen
-
-    Returns:
-       tuple: fastq_screen output names (without leading path)
-
-    """
-    base_name = "%s_%s_screen" % (strip_ngs_extensions(os.path.basename(fastq)),
-                                  str(screen_name))
-    
-    return (base_name+'.png',base_name+'.txt')
-
-def fastqc_output(fastq):
-    """
-    Generate name of FastQC outputs
-
-    Given a Fastq file name, the outputs from FastQC will look
-    like:
-
-    - {FASTQ}_fastqc/
-    - {FASTQ}_fastqc.html
-    - {FASTQ}_fastqc.zip
-
-    Arguments:
-       fastq (str): name of Fastq file
-
-    Returns:
-       tuple: FastQC outputs (without leading paths)
-
-    """
-    base_name = "%s_fastqc" % strip_ngs_extensions(os.path.basename(fastq))
-    return (base_name,base_name+'.html',base_name+'.zip')
-
-def fastq_strand_output(fastq):
-    """
-    Generate name for fastq_strand.py output
-
-    Given a Fastq file name, the output from fastq_strand.py
-    will look like:
-
-    - {FASTQ}_fastq_strand.txt
-
-    Arguments:
-       fastq (str): name of Fastq file
-
-    Returns:
-       tuple: fastq_strand.py output (without leading paths)
-
-    """
-    return "%s_fastq_strand.txt" % strip_ngs_extensions(
-        os.path.basename(fastq))
