--- auto_process_ngs/test/qc/test_illumina_qc.py
@@ -1,63 +1,0 @@
-#######################################################################
-# Unit tests for qc/illumina_qc.py
-#######################################################################
-
-import unittest
-import tempfile
-import shutil
-import os
-from auto_process_ngs.mock import MockIlluminaQcSh
-from auto_process_ngs.mock import MockAnalysisProject
-from auto_process_ngs.analysis import AnalysisProject
-from auto_process_ngs.qc.illumina_qc import fastq_screen_output
-from auto_process_ngs.qc.illumina_qc import fastqc_output
-from auto_process_ngs.qc.illumina_qc import fastq_strand_output
-
-# Set to False to keep test output dirs
-REMOVE_TEST_OUTPUTS = True
-
-class TestFastqScreenOutputFunction(unittest.TestCase):
-    def test_fastq_screen_output(self):
-        """fastq_screen_output: handles .fastq file
-        """
-        self.assertEqual(fastq_screen_output('/data/PB/PB1_ATTAGG_L001_R1_001.fastq',
-                                             'model_organisms'),
-                         ('PB1_ATTAGG_L001_R1_001_model_organisms_screen.png',
-                          'PB1_ATTAGG_L001_R1_001_model_organisms_screen.txt'))
-    def test_fastq_screen_output_fastqgz(self):
-        """fastq_screen_output: handles fastq.gz file
-        """
-        self.assertEqual(fastq_screen_output('/data/PB/PB1_ATTAGG_L001_R1_001.fastq.gz',
-                                             'model_organisms'),
-                         ('PB1_ATTAGG_L001_R1_001_model_organisms_screen.png',
-                          'PB1_ATTAGG_L001_R1_001_model_organisms_screen.txt'))
-
-class TestFastqcOutputFunction(unittest.TestCase):
-    def test_fastqc_output(self):
-        """fastqc_output: handles .fastq file
-        """
-        self.assertEqual(fastqc_output('/data/PB/PB1_ATTAGG_L001_R1_001.fastq'),
-                         ('PB1_ATTAGG_L001_R1_001_fastqc',
-                          'PB1_ATTAGG_L001_R1_001_fastqc.html',
-                          'PB1_ATTAGG_L001_R1_001_fastqc.zip'))
-    def test_fastqc_output_fastqgz(self):
-        """fastqc_output: handles fastq.gz file
-        """
-        self.assertEqual(fastqc_output('/data/PB/PB1_ATTAGG_L001_R1_001.fastq.gz'),
-                         ('PB1_ATTAGG_L001_R1_001_fastqc',
-                          'PB1_ATTAGG_L001_R1_001_fastqc.html',
-                          'PB1_ATTAGG_L001_R1_001_fastqc.zip'))
-
-class TestFastqStrandOutputFunction(unittest.TestCase):
-    def test_fastq_strand_output(self):
-        """fastq_strand_output: handles .fastq file
-        """
-        self.assertEqual(fastq_strand_output(
-            '/data/PB/PB1_ATTAGG_L001_R1_001.fastq'),
-                         'PB1_ATTAGG_L001_R1_001_fastq_strand.txt')
-    def test_fastq_strand_output_fastqgz(self):
-        """fastq_strand_output: handles fastq.gz file
-        """
-        self.assertEqual(fastq_strand_output(
-            '/data/PB/PB1_ATTAGG_L001_R1_001.fastq.gz'),
-                         'PB1_ATTAGG_L001_R1_001_fastq_strand.txt')
