+++ auto_process_ngs/test/qc/test_pipeline.py
@@ -17,7 +17,6 @@ from auto_process_ngs.qc.pipeline import QCPipeline
 from auto_process_ngs.qc.pipeline import copy_project
 from auto_process_ngs.qc.pipeline import check_illumina_qc_outputs
 from auto_process_ngs.qc.pipeline import check_fastq_strand_outputs
-from auto_process_ngs.qc.pipeline import expected_outputs
 
 # Set to False to keep test output dirs
 REMOVE_TEST_OUTPUTS = True
@@ -949,254 +948,3 @@ class TestCheckFastqStrandOutputs(unittest.TestCase):
                                                     fastq_strand_conf,
                                                     qc_protocol="singlecell"),
                          [])
-
-class TestExpectedOutputs(unittest.TestCase):
-    """
-    Tests for the 'expected_outputs' function
-    """
-    def setUp(self):
-        # Create a temp working dir
-        self.wd = tempfile.mkdtemp(suffix='TestExpectedOutputs')
-
-    def tearDown(self):
-        # Remove the temporary test directory
-        if REMOVE_TEST_OUTPUTS:
-            shutil.rmtree(self.wd)
-
-    def test_expected_outputs_standardPE(self):
-        """
-        expected_outputs: standard paired-end, no strandedness
-        """
-        # Make mock analysis project
-        p = MockAnalysisProject("PJB",("PJB1_S1_R1_001.fastq.gz",
-                                       "PJB1_S1_R2_001.fastq.gz",),
-                                metadata={ 'Organism': 'Human' })
-        p.create(top_dir=self.wd)
-        # Reference outputs
-        reference_outputs = ("PJB1_S1_R1_001_fastqc",
-                             "PJB1_S1_R1_001_fastqc.html",
-                             "PJB1_S1_R1_001_fastqc.zip",
-                             "PJB1_S1_R1_001_model_organisms_screen.png",
-                             "PJB1_S1_R1_001_model_organisms_screen.txt",
-                             "PJB1_S1_R1_001_other_organisms_screen.png",
-                             "PJB1_S1_R1_001_other_organisms_screen.txt",
-                             "PJB1_S1_R1_001_rRNA_screen.png",
-                             "PJB1_S1_R1_001_rRNA_screen.txt",
-                             "PJB1_S1_R2_001_fastqc",
-                             "PJB1_S1_R2_001_fastqc.html",
-                             "PJB1_S1_R2_001_fastqc.zip",
-                             "PJB1_S1_R2_001_model_organisms_screen.png",
-                             "PJB1_S1_R2_001_model_organisms_screen.txt",
-                             "PJB1_S1_R2_001_other_organisms_screen.png",
-                             "PJB1_S1_R2_001_other_organisms_screen.txt",
-                             "PJB1_S1_R2_001_rRNA_screen.png",
-                             "PJB1_S1_R2_001_rRNA_screen.txt",)
-        expected = expected_outputs(AnalysisProject(p.name,
-                                                    os.path.join(self.wd,
-                                                                 p.name)),
-                                    "qc",
-                                    qc_protocol="standardPE")
-        for e in expected:
-            self.assertEqual(os.path.dirname(e),os.path.join(self.wd,
-                                                             p.name,
-                                                             "qc"))
-            self.assertTrue(os.path.basename(e) in reference_outputs)
-        for r in reference_outputs:
-            self.assertTrue(os.path.join(self.wd,p.name,"qc",r) in expected)
-
-    def test_expected_outputs_standardPE_with_strand(self):
-        """
-        expected_outputs: standard paired-end with strandedness
-        """
-        # Make mock analysis project
-        p = MockAnalysisProject("PJB",("PJB1_S1_R1_001.fastq.gz",
-                                       "PJB1_S1_R2_001.fastq.gz",),
-                                metadata={ 'Organism': 'Human' })
-        p.create(top_dir=self.wd)
-        # Make mock fastq_strand
-        mock_fastq_strand_conf = os.path.join(self.wd,
-                                              p.name,
-                                              "fastq_strand.conf")
-        with open(mock_fastq_strand_conf,'w') as fp:
-            fp.write("")
-        # Reference outputs
-        reference_outputs = ("PJB1_S1_R1_001_fastqc",
-                             "PJB1_S1_R1_001_fastqc.html",
-                             "PJB1_S1_R1_001_fastqc.zip",
-                             "PJB1_S1_R1_001_model_organisms_screen.png",
-                             "PJB1_S1_R1_001_model_organisms_screen.txt",
-                             "PJB1_S1_R1_001_other_organisms_screen.png",
-                             "PJB1_S1_R1_001_other_organisms_screen.txt",
-                             "PJB1_S1_R1_001_rRNA_screen.png",
-                             "PJB1_S1_R1_001_rRNA_screen.txt",
-                             "PJB1_S1_R2_001_fastqc",
-                             "PJB1_S1_R2_001_fastqc.html",
-                             "PJB1_S1_R2_001_fastqc.zip",
-                             "PJB1_S1_R2_001_model_organisms_screen.png",
-                             "PJB1_S1_R2_001_model_organisms_screen.txt",
-                             "PJB1_S1_R2_001_other_organisms_screen.png",
-                             "PJB1_S1_R2_001_other_organisms_screen.txt",
-                             "PJB1_S1_R2_001_rRNA_screen.png",
-                             "PJB1_S1_R2_001_rRNA_screen.txt",
-                             "PJB1_S1_R1_001_fastq_strand.txt",)
-        expected = expected_outputs(AnalysisProject(p.name,
-                                                    os.path.join(self.wd,
-                                                                 p.name)),
-                                    "qc",
-                                    fastq_strand_conf=mock_fastq_strand_conf,
-                                    qc_protocol="standardPE")
-        for e in expected:
-            self.assertEqual(os.path.dirname(e),os.path.join(self.wd,
-                                                             p.name,
-                                                             "qc"))
-            self.assertTrue(os.path.basename(e) in reference_outputs)
-        for r in reference_outputs:
-            self.assertTrue(os.path.join(self.wd,p.name,"qc",r) in expected)
-
-    def test_expected_outputs_standardSE(self):
-        """
-        expected_outputs: standard single-end, no strandedness
-        """
-        # Make mock analysis project
-        p = MockAnalysisProject("PJB",("PJB1_S1_R1_001.fastq.gz",),
-                                metadata={ 'Organism': 'Human' })
-        p.create(top_dir=self.wd)
-        # Reference outputs
-        reference_outputs = ("PJB1_S1_R1_001_fastqc",
-                             "PJB1_S1_R1_001_fastqc.html",
-                             "PJB1_S1_R1_001_fastqc.zip",
-                             "PJB1_S1_R1_001_model_organisms_screen.png",
-                             "PJB1_S1_R1_001_model_organisms_screen.txt",
-                             "PJB1_S1_R1_001_other_organisms_screen.png",
-                             "PJB1_S1_R1_001_other_organisms_screen.txt",
-                             "PJB1_S1_R1_001_rRNA_screen.png",
-                             "PJB1_S1_R1_001_rRNA_screen.txt",)
-        expected = expected_outputs(AnalysisProject(p.name,
-                                                    os.path.join(self.wd,
-                                                                 p.name)),
-                                    "qc",
-                                    qc_protocol="standardSE")
-        for e in expected:
-            self.assertEqual(os.path.dirname(e),os.path.join(self.wd,
-                                                             p.name,
-                                                             "qc"))
-            self.assertTrue(os.path.basename(e) in reference_outputs)
-        for r in reference_outputs:
-            self.assertTrue(os.path.join(self.wd,p.name,"qc",r) in expected)
-
-    def test_expected_outputs_standardSE_with_strand(self):
-        """
-        expected_outputs: standard single-end with strandedness
-        """
-        # Make mock analysis project
-        p = MockAnalysisProject("PJB",("PJB1_S1_R1_001.fastq.gz",),
-                                metadata={ 'Organism': 'Human' })
-        p.create(top_dir=self.wd)
-        # Make mock fastq_strand
-        mock_fastq_strand_conf = os.path.join(self.wd,
-                                              p.name,
-                                              "fastq_strand.conf")
-        with open(mock_fastq_strand_conf,'w') as fp:
-            fp.write("")
-        # Reference outputs
-        reference_outputs = ("PJB1_S1_R1_001_fastqc",
-                             "PJB1_S1_R1_001_fastqc.html",
-                             "PJB1_S1_R1_001_fastqc.zip",
-                             "PJB1_S1_R1_001_model_organisms_screen.png",
-                             "PJB1_S1_R1_001_model_organisms_screen.txt",
-                             "PJB1_S1_R1_001_other_organisms_screen.png",
-                             "PJB1_S1_R1_001_other_organisms_screen.txt",
-                             "PJB1_S1_R1_001_rRNA_screen.png",
-                             "PJB1_S1_R1_001_rRNA_screen.txt",
-                             "PJB1_S1_R1_001_fastq_strand.txt",)
-        expected = expected_outputs(AnalysisProject(p.name,
-                                                    os.path.join(self.wd,
-                                                                 p.name)),
-                                    "qc",
-                                    fastq_strand_conf=mock_fastq_strand_conf,
-                                    qc_protocol="standardSE")
-        for e in expected:
-            self.assertEqual(os.path.dirname(e),os.path.join(self.wd,
-                                                             p.name,
-                                                             "qc"))
-            self.assertTrue(os.path.basename(e) in reference_outputs)
-        for r in reference_outputs:
-            self.assertTrue(os.path.join(self.wd,p.name,"qc",r) in expected)
-
-    def test_expected_outputs_singlecell(self):
-        """
-        expected_outputs: single-cell, no strandedness
-        """
-        # Make mock analysis project
-        p = MockAnalysisProject("PJB",("PJB1_S1_R1_001.fastq.gz",
-                                       "PJB1_S1_R2_001.fastq.gz",),
-                                metadata={ 'Organism': 'Human' })
-        p.create(top_dir=self.wd)
-        # Reference outputs
-        reference_outputs = ("PJB1_S1_R1_001_fastqc",
-                             "PJB1_S1_R1_001_fastqc.html",
-                             "PJB1_S1_R1_001_fastqc.zip",
-                             "PJB1_S1_R2_001_fastqc",
-                             "PJB1_S1_R2_001_fastqc.html",
-                             "PJB1_S1_R2_001_fastqc.zip",
-                             "PJB1_S1_R2_001_model_organisms_screen.png",
-                             "PJB1_S1_R2_001_model_organisms_screen.txt",
-                             "PJB1_S1_R2_001_other_organisms_screen.png",
-                             "PJB1_S1_R2_001_other_organisms_screen.txt",
-                             "PJB1_S1_R2_001_rRNA_screen.png",
-                             "PJB1_S1_R2_001_rRNA_screen.txt",)
-        expected = expected_outputs(AnalysisProject(p.name,
-                                                    os.path.join(self.wd,
-                                                                 p.name)),
-                                    "qc",
-                                    qc_protocol="singlecell")
-        for e in expected:
-            self.assertEqual(os.path.dirname(e),os.path.join(self.wd,
-                                                             p.name,
-                                                             "qc"))
-            self.assertTrue(os.path.basename(e) in reference_outputs)
-        for r in reference_outputs:
-            self.assertTrue(os.path.join(self.wd,p.name,"qc",r) in expected)
-
-    def test_expected_outputs_singlecell_with_strand(self):
-        """
-        expected_outputs: single-cell with strandedness
-        """
-        # Make mock analysis project
-        p = MockAnalysisProject("PJB",("PJB1_S1_R1_001.fastq.gz",
-                                       "PJB1_S1_R2_001.fastq.gz",),
-                                metadata={ 'Organism': 'Human' })
-        p.create(top_dir=self.wd)
-        # Make mock fastq_strand
-        mock_fastq_strand_conf = os.path.join(self.wd,
-                                              p.name,
-                                              "fastq_strand.conf")
-        with open(mock_fastq_strand_conf,'w') as fp:
-            fp.write("")
-        # Reference outputs
-        reference_outputs = ("PJB1_S1_R1_001_fastqc",
-                             "PJB1_S1_R1_001_fastqc.html",
-                             "PJB1_S1_R1_001_fastqc.zip",
-                             "PJB1_S1_R2_001_fastqc",
-                             "PJB1_S1_R2_001_fastqc.html",
-                             "PJB1_S1_R2_001_fastqc.zip",
-                             "PJB1_S1_R2_001_model_organisms_screen.png",
-                             "PJB1_S1_R2_001_model_organisms_screen.txt",
-                             "PJB1_S1_R2_001_other_organisms_screen.png",
-                             "PJB1_S1_R2_001_other_organisms_screen.txt",
-                             "PJB1_S1_R2_001_rRNA_screen.png",
-                             "PJB1_S1_R2_001_rRNA_screen.txt",
-                             "PJB1_S1_R2_001_fastq_strand.txt",)
-        expected = expected_outputs(AnalysisProject(p.name,
-                                                    os.path.join(self.wd,
-                                                                 p.name)),
-                                    "qc",
-                                    fastq_strand_conf=mock_fastq_strand_conf,
-                                    qc_protocol="singlecell")
-        for e in expected:
-            self.assertEqual(os.path.dirname(e),os.path.join(self.wd,
-                                                             p.name,
-                                                             "qc"))
-            self.assertTrue(os.path.basename(e) in reference_outputs)
-        for r in reference_outputs:
-            self.assertTrue(os.path.join(self.wd,p.name,"qc",r) in expected)
